webapp/requirements.txt
streamlit
openai
python-dotenv
duckdb
Pillow
requests
